feat(smart): enforce HTTPS on registration_endpoint; extract URIValidation module
